feat: 남은 UserRepository도 spring data jpa로 교체 #5
